Archive for a scratched attempt at typing
There are still some great ideas in there, so I didn't want to completely remove everything and instead "archive" it in this branch.

This branch contains a `Type` type that is a big struct with different fields for various types. It sounded great on paper, but...

- It has a very big footprint, even though usually an actual type will only use a few types. So it always comes with a lot of wasted space.
  - There was an attempt to work around this, but it introduced more unnecessary complexity...
- At first, it felt like I could decouple different types, but in the end it was just more hassle than it was worth it...

Overall it still felt like this whole venture helped my understanding of the whole thing quite a bit though.